pyproject.toml
[tool.poetry]
name = "arxivchat"
version = "0.1.0"
description = ""
authors = ["Mateusz Kozak <mateusz@mkozak.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
llama-index = "^0.5.9"
uvicorn = "^0.21.1"
fastapi = "^0.95.0"
pypdf = "^3.7.0"
chromadb = "^0.3.21"
arxiv = "^1.4.4"
python-dotenv = "^1.0.0"
sentry-sdk = "^1.19.1"
celery = {extras = ["redis"], version = "^5.2.7"}
sentence-transformers = "^2.2.2"
qdrant-client = "^1.1.5"
langchain = {extras = ["all"], version = ">0.0.148"}
feedparser = "^6.0.10"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "~23.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"